requirements.txt
kafka-python==2.0.2
pyspark==3.4.0
pandas==2.0.3
faker==19.3.0